# ipx/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Solana transaction API",
        "title": "SVMlab API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:33153",
    "basePath": "/",
    "paths": {
...
        "/svm/cluster/transaction/fee": {
            "post": {
                "description": "The fee follows from the signature count and any compute budget instructions, both of which live in the message, so it is known before signing and cannot be exceeded by execution the way a gas estimate can.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cluster"
                ],
                "summary": "Price a message",
                "parameters": [
                    {
                        "description": "Message",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/misc.FeeRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Chain name, e.g. solana",
                        "name": "X-Chain-Name",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Chain network, e.g. testnet",
                        "name": "X-Chain-Network",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/misc.FeeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/svm/cluster/transaction/send": {
            "post": {
                "description": "Submits a fully signed transaction to the cluster and returns its signature. Acceptance is not execution; the transaction still has to land in a block.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cluster"
                ],
                "summary": "Broadcast a signed transaction",
                "parameters": [
                    {
                        "description": "Signed transaction",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/misc.SendTransactionRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Chain name, e.g. solana",
                        "name": "X-Chain-Name",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Chain network, e.g. testnet",
                        "name": "X-Chain-Network",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/misc.SendTransactionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/svm/cluster/transaction/simulate": {
            "post": {
                "description": "Runs a fully signed transaction against the node's state without broadcasting it, and returns the program logs either way. The logs are the only account of why execution stopped; nothing here corresponds to a revert string.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cluster"
                ],
                "summary": "Execute a transaction without submitting it",
                "parameters": [
                    {
                        "description": "Transaction",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/misc.SimulateTransactionRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Chain name, e.g. solana",
                        "name": "X-Chain-Name",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Chain network, e.g. testnet",
                        "name": "X-Chain-Network",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/misc.SimulateTransactionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/svm/cluster/transaction/status": {
            "post": {
                "description": "Stands in for an EVM receipt, but found can stay false forever: a transaction whose blockhash expired is forgotten and never lands.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cluster"
                ],
                "summary": "Read a transaction's status",
                "parameters": [
                    {
                        "description": "Signature",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/misc.SignatureStatusRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Chain name, e.g. solana",
                        "name": "X-Chain-Name",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Chain network, e.g. testnet",
                        "name": "X-Chain-Network",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/misc.SignatureStatusResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "misc.FeeRequest": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "misc.FeeResponse": {
            "type": "object",
            "properties": {
                "lamports": {
                    "type": "string"
                },
                "sol": {
                    "type": "string"
                },
                "valid": {
                    "type": "boolean"
                }
            }
        },
...
        "misc.SendTransactionRequest": {
            "type": "object",
            "properties": {
                "transaction": {
                    "type": "string"
                }
            }
        },
        "misc.SendTransactionResponse": {
            "type": "object",
            "properties": {
                "signature": {
                    "type": "string"
                }
            }
        },
...
        "misc.SignatureStatusRequest": {
            "type": "object",
            "properties": {
                "signature": {
                    "type": "string"
                }
            }
        },
        "misc.SignatureStatusResponse": {
            "type": "object",
            "properties": {
                "confirmation_status": {
                    "type": "string"
                },
                "confirmations": {
                    "type": "integer"
                },
                "err": {
                    "type": "string"
                },
                "failed": {
                    "type": "boolean"
                },
                "found": {
                    "type": "boolean"
                },
                "signature": {
                    "type": "string"
                },
                "slot": {
                    "type": "integer"
                }
            }
        },
        "misc.SimulateTransactionRequest": {
            "type": "object",
            "properties": {
                "transaction": {
                    "type": "string"
                }
            }
        },
        "misc.SimulateTransactionResponse": {
            "type": "object",
            "properties": {
                "err": {
                    "type": "string"
                },
                "failed": {
                    "type": "boolean"
                },
                "logs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "units_consumed": {
                    "type": "integer"
                }
            }
        },
...
    }
}